# Improve performance of `read_ine_file`