--- a/pydrad/parse/tests/conftest.py
+++ b/pydrad/parse/tests/conftest.py
@@ -1,0 +1,11 @@
+"""
+Common fixtures for testing pydrad.parse
+"""
+import pytest
+
+from pydrad.parse import Strand
+
+
+@pytest.fixture
+def strand(hydrad):
+    return Strand(hydrad)