--- a/pydrad/parse/tests/test_strand.py
+++ b/pydrad/parse/tests/test_strand.py
@@ -4,6 +4,7 @@ Test parsing HYDRAD results
 import astropy.units as u
 import h5py
 import numpy as np
+import plasmapy.particles
 import pytest
 
 from pydrad.parse import Profile, Strand
@@ -78,6 +79,16 @@ def strand_only_amr(hydrad):
     return Strand(hydrad,
                   read_phy=False,
                   read_ine=False,
+                  read_trm=False,
+                  read_hstate=False,
+                  read_scl=False)
+
+
+@pytest.fixture
+def strand_ine(hydrad):
+    return Strand(hydrad,
+                  read_phy=False,
+                  read_ine=True,
                   read_trm=False,
                   read_hstate=False,
                   read_scl=False)
@@ -206,3 +217,13 @@ def test_profile_instantiation(strand_only_amr, strand_only_amr_time_cfg):
     # No index, master time
     p = Profile(strand_only_amr.hydrad_root, strand_only_amr.time[1], master_time=strand_only_amr._master_time)
     assert p._index == 1
+
+
+def test_ine_results(strand_ine):
+    for profile in strand_ine:
+        for element in strand_ine.config['radiation']['elements_nonequilibrium']:
+            Z = plasmapy.particles.atomic_number(element)
+            for i_z in range(1,Z+2):
+                assert hasattr(profile, f'{element}_{i_z}')
+                ion_frac = getattr(profile, f'{element}_{i_z}')
+                assert ion_frac.shape == profile.coordinate.shape