Simplifies and speeds up parsing of `.ine` files.

Before this PR

```python
%%timeit
_ = read_ine_file(ine_filename, n_s)
23.1 ms ± 439 μs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

After this PR

```python
%%timeit
_ = read_ine_file(ine_filename, n_s)
13.2 ms ± 481 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

10 ms might not seem like much, but can make a significant difference when parsing many of these files. For example, for 6000 files, this PR makes the parsing of all of these files 1 minute faster.